microUser/requirements.txt
fastapi==0.115
uvicorn==0.34.0
python-decouple==3.8
asyncpg==0.30.0
loguru==0.7.3
httpx==0.28.1
passlib==1.7.4
bcrypt==3.2.0
python-jose==3.3.0
requests==2.27.1
python-multipart==0.0.5